Reorganize code: new class HardwareCollector to store all data. Initialize once and then just read!
